requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
pickle-mixin

langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-openai
